CORE-13858 - Record inbound message rate in the loopback case
