+++ components/link-manager/src/main/kotlin/net/corda/p2p/linkmanager/inbound/InboundMessageProcessor.kt
@@ -32,6 +32,7 @@ import net.corda.p2p.linkmanager.sessions.SessionManager
 import net.corda.data.p2p.markers.AppMessageMarker
 import net.corda.data.p2p.markers.LinkManagerReceivedMarker
 import net.corda.metrics.CordaMetrics
+import net.corda.p2p.linkmanager.metrics.recordInboundMessagesMetric
 import net.corda.schema.Schemas
 import net.corda.utilities.debug
 import net.corda.utilities.time.Clock
@@ -286,35 +287,6 @@ internal class InboundMessageProcessor(
             message.messageId,
             AppMessageMarker(LinkManagerReceivedMarker(), clock.instant().toEpochMilli())
         )
-    }
-
-    private fun recordInboundMessagesMetric(message: AuthenticatedMessage) {
-        message.header.let {
-            recordInboundMessagesMetric(it.source.x500Name, it.destination.x500Name, it.source.groupId,
-                it.subsystem, message::class.java.simpleName)
-        }
-    }
-
-    private fun recordInboundMessagesMetric(message: InboundUnauthenticatedMessage) {
-        recordInboundMessagesMetric(null, null, null,
-            message.header.subsystem, message::class.java.simpleName)
-    }
-
-    private fun recordInboundMessagesMetric(source: String?, dest: String?, group: String?, subsystem: String, messageType: String) {
-        val builder = CordaMetrics.Metric.InboundMessageCount.builder()
-        listOf(
-            CordaMetrics.Tag.SourceVirtualNode to source,
-            CordaMetrics.Tag.DestinationVirtualNode to dest,
-            CordaMetrics.Tag.MembershipGroup to group,
-            CordaMetrics.Tag.MessagingSubsystem to subsystem,
-            CordaMetrics.Tag.MessageType to messageType,
-        ).forEach {
-            val value = it.second
-            if (value != null) {
-                builder.withTag(it.first, value)
-            }
-        }
-        builder.build().increment()
     }
 
     private fun <T> checkAllowedCommunication(
